Dumb idea: fs should be class property
The advantage of this is that most pathlib.Path methods that return Path
objects don't have to be re-implemented, because they use the object's
class to construct new paths, which works just fine with this approach
(same constructor signature and all).

But it's kind of stupid because fs is really not a matter of the
class...

This whole idea was born out of pure laziness.